Move `react-router-dom` as an explicit dependency
While `react-router-dom` is not always required at runtime (only really requested when using `<SecureRoute />`), having it defined as a peer dependency raise issues for any user wanted to move to next majors of `react-router-dom` and not relying on `<SecureRoute />`.

Indeed for such users of `okta-react`, WebPack and other bundlers will fail to build the project because of the import of `useMatchRoute` not being part of the new major of `react-router-dom`. While it probably make a bit of sense whne you use `<SecureRoute />`, having your build broken because of it while you don't need it at all is strange. By moving the library as a clear dependency, `okta-react` wil always pull the right version and stop making user's builds fail.